# Conflicts:
#	src/main/java/ru/skillbox/socialnetwork/services/PostService.java

+++ src/main/java/ru/skillbox/socialnetwork/controllers/PostController.java
@@ -42,40 +42,40 @@ public class PostController {
 
 
   @PutMapping("/{id}")
-  public ResponseEntity<?> putApiPostId(@PathVariable("id") int id, @Param("publish_date") long publishDate,
+  public ResponseEntity<?> putApiPostId(@PathVariable("id") long id, @Param("publish_date") long publishDate,
                                         @RequestBody TitlePostTextRequest requestBody) {
     return postService.putApiPostId(id, publishDate, requestBody);
   }
 
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<?> deleteApiPostId(@PathVariable("id") int id) {
+  public ResponseEntity<?> deleteApiPostId(@PathVariable("id") long id) {
     return postService.deleteApiPostId(id);
   }
 
 
   @PutMapping("/{id}/recover")
-  public ResponseEntity<?> putApiPostIdRecover(@PathVariable("id") int id) {
+  public ResponseEntity<?> putApiPostIdRecover(@PathVariable("id") long id) {
     return postService.putApiPostIdRecover(id);
   }
 
 
   @GetMapping("/{id}/comments")
-  public ResponseEntity<?> getApiPostIdComments(@PathVariable("id") int id, @Param("offset") int offset,
+  public ResponseEntity<?> getApiPostIdComments(@PathVariable("id") long id, @Param("offset") int offset,
                                                 @Param("itemPerPage") int itemPerPage) {
     return postService.getApiPostIdComments(id, offset, itemPerPage);
   }
 
 
   @PostMapping("/{id}/comments")
-  public ResponseEntity<?> postApiPostIdComments(@PathVariable("id") int id,
+  public ResponseEntity<?> postApiPostIdComments(@PathVariable("id") long id,
                                                  @RequestBody ParentIdCommentTextRequest requestBody) {
     return postService.postApiPostIdComments(id, requestBody);
   }
 
 
   @PutMapping("/{id}/comments/{comment_id}")
-  public ResponseEntity<?> putApiPostIdCommentsCommentId(@PathVariable("id") int id,
+  public ResponseEntity<?> putApiPostIdCommentsCommentId(@PathVariable("id") long id,
                                                          @PathVariable("comment_id") int commentId,
                                                          @RequestBody ParentIdCommentTextRequest requestBody) {
     return postService.putApiPostIdCommentsCommentId(id, commentId, requestBody);
@@ -83,27 +83,27 @@ public class PostController {
 
 
   @DeleteMapping("/{id}/comments/{comment_id}")
-  public ResponseEntity<?> deleteApiPostIdCommentsCommentId(@PathVariable("id") int id,
+  public ResponseEntity<?> deleteApiPostIdCommentsCommentId(@PathVariable("id") long id,
                                                             @PathVariable("comment_id") int commentId) {
     return postService.deleteApiPostIdCommentsCommentId(id, commentId);
   }
 
 
   @PutMapping("/{id}/comments/{comment_id}/recover")
-  public ResponseEntity<?> putApiPostIdCommentsCommentId(@PathVariable("id") int id,
+  public ResponseEntity<?> putApiPostIdCommentsCommentId(@PathVariable("id") long id,
                                                          @PathVariable("comment_id") int commentId) {
     return postService.putApiPostIdCommentsCommentId(id, commentId);
   }
 
 
   @PostMapping("/{id}/report")
-  public ResponseEntity<?> postApiPostIdReport(@PathVariable("id") int id) {
+  public ResponseEntity<?> postApiPostIdReport(@PathVariable("id") long id) {
     return postService.postApiPostIdReport(id);
   }
 
 
   @PostMapping("/{id}/comments/{comment_id}/report")
-  public ResponseEntity<?> postApiPostIdCommentsCommentIdReport(@PathVariable("id") int id,
+  public ResponseEntity<?> postApiPostIdCommentsCommentIdReport(@PathVariable("id") long id,
                                                                 @PathVariable("comment_id") int commentId) {
     return postService.postApiPostIdCommentsCommentIdReport(id, commentId);
   }
